fix mistake for illegal set fieldtype when setLong val for issue https://github.com/datavane/tis/issues/267

+++ chunjun-connectors/chunjun-connector-jdbc-base/src/main/java/com/dtstack/chunjun/connector/jdbc/sink/IFieldNamesAttachedStatement.java
@@ -44,7 +44,7 @@ public interface IFieldNamesAttachedStatement {
     }
 
     default void setLong(int pos, Long val) throws SQLException {
-        getFieldNamedPstmt().setDouble(pos, val);
+        getFieldNamedPstmt().setLong(pos, val);
     }
 
     default void setBigDecimal(int pos, BigDecimal val) throws SQLException {
